+++ src/LineChart/LineChart.story.tsx
@@ -1,54 +1,27 @@
-import { storiesOf } from '@storybook/react';
-import chroma from 'chroma-js';
 import { timeDay } from 'd3-time';
 import React, { Fragment, useState } from 'react';
 import moment from 'moment';
 import { object, number, select } from '@storybook/addon-knobs';
-
 import {
-  multiDateData,
   singleDateData,
   largeDateData,
-  randomNumber,
-  longMultiDateData
+  randomNumber
 } from '../../demo';
 import { LineChart, LineSeries } from './LineChart';
 import {
-  StackedAreaChart,
-  StackedNormalizedAreaChart,
-  StackedAreaSeries,
-  Line,
-  StackedNormalizedAreaSeries
+  Line
 } from '../AreaChart';
-import { GridlineSeries, Gridline } from '../common/Gridline';
-import { PointSeries } from '../AreaChart';
 import {
   LinearXAxisTickSeries,
-  LinearXAxis,
-  LinearYAxisTickSeries,
-  LinearYAxis
+  LinearXAxis
 } from '../common/Axis/LinearAxis';
-import { ScatterPoint } from '../ScatterPlot';
-import { symbol, symbolStar } from 'd3-shape';
 import { schemes } from '../common/color';
 
-const prettyData = (() => {
-  const data: any[] = [];
-  for (let i = 0; i < 20; i++) {
-    const series: any[] = [];
-    for (let j = 0; j < 100; j++) {
-      series.push({
-        key: j,
-        data: (i / 10 + 1) * Math.sin((Math.PI * (i + j)) / 50)
-      });
-    }
-    data.push({ key: i, data: series });
-  }
-  return data;
-})();
+export default {
+  title: 'Charts/Line Chart/Single Series',
+};
 
-storiesOf('Charts/Line Chart/Single Series', module)
-  .add('Simple', () => {
+export const Simple = () => {
     const height = number('Height', 250);
     const width = number('Width', 350);
     const lineStroke = number('Stroke Width', 4);
@@ -78,21 +51,24 @@ storiesOf('Charts/Line Chart/Single Series', module)
         }
       />
     );
-  })
-  .add('No Animation', () => (
+  };
+
+export const NoAnimation = () => (
     <LineChart
       width={350}
       height={250}
       data={singleDateData}
       series={<LineSeries animated={false} />}
     />
-  ))
-  .add('Autosize', () => (
+  );
+
+export const Autosize = () => (
     <div style={{ width: '50vw', height: '50vh', border: 'solid 1px red' }}>
       <LineChart data={singleDateData} xAxis={<LinearXAxis type="time" />} />
     </div>
-  ))
-  .add('Interval', () => (
+  );
+
+export const Interval = () => (
     <LineChart
       width={350}
       height={250}
@@ -104,8 +80,9 @@ storiesOf('Charts/Line Chart/Single Series', module)
         />
       }
     />
-  ))
-  .add('Large Dataset', () => {
+  );
+
+export const LargeDataset = () => {
     const height = number('Height', 250);
     const width = number('Width', 400);
     const color = select('Color Scheme', schemes, 'cybertron');
@@ -131,8 +108,9 @@ storiesOf('Charts/Line Chart/Single Series', module)
         xAxis={<LinearXAxis type="time" />}
       />
     );
-  })
-  .add('Dynamic Colors', () => (
+  };
+
+export const DynamicColors = () => (
     <LineChart
       width={350}
       height={250}
@@ -143,256 +121,9 @@ storiesOf('Charts/Line Chart/Single Series', module)
         />
       }
     />
-  ))
-  .add('Live Updating', () => <LiveUpdatingStory />);
+  );
 
-storiesOf('Charts/Line Chart/Multi Series', module)
-  .add('Simple', () => {
-    const height = number('Height', 250);
-    const width = number('Width', 550);
-    const lineStroke = number('Stroke Width', 4);
-    const color = select('Color Scheme', schemes, 'cybertron');
-    const data = object('Data', multiDateData);
-
-    return (
-      <LineChart
-        width={width}
-        height={height}
-        series={
-          <LineSeries
-            type="grouped"
-            line={<Line strokeWidth={lineStroke} />}
-            colorScheme={color}
-          />
-        }
-        data={data}
-      />
-    );
-  })
-  .add('Large Data', () => (
-    <LineChart
-      width={400}
-      height={300}
-      data={prettyData}
-      yAxis={
-        <LinearYAxis
-          scaled={true}
-          type="value"
-          axisLine={null}
-          tickSeries={<LinearYAxisTickSeries line={null} label={null} />}
-        />
-      }
-      xAxis={
-        <LinearXAxis
-          type="value"
-          scaled={true}
-          axisLine={null}
-          tickSeries={<LinearXAxisTickSeries line={null} label={null} />}
-        />
-      }
-      series={
-        <LineSeries
-          type="grouped"
-          line={<Line strokeWidth={1} />}
-          colorScheme="cybertron"
-        />
-      }
-    />
-  ))
-  .add('Custom Line Styles', () => (
-    <LineChart
-      width={550}
-      height={350}
-      series={
-        <LineSeries
-          type="grouped"
-          line={
-            <Line
-              strokeWidth={3}
-              style={(data) => {
-                if (
-                  data &&
-                  data.length &&
-                  data[0] &&
-                  data[0].key === 'Threat Intel'
-                ) {
-                  console.log('Style callback...', data);
-                  return {
-                    strokeDasharray: '5'
-                  };
-                }
-              }}
-            />
-          }
-          colorScheme={chroma
-            .scale(['27efb5', '00bfff'])
-            .colors(multiDateData.length)}
-        />
-      }
-      data={multiDateData}
-    />
-  ))
-  .add('Large Dataset', () => (
-    <LineChart
-      width={550}
-      height={350}
-      series={
-        <LineSeries
-          type="grouped"
-          colorScheme={chroma
-            .scale(['ACB7C9', '418AD7'])
-            .colors(longMultiDateData.length)}
-        />
-      }
-      data={longMultiDateData}
-    />
-  ))
-  .add('Stacked', () => {
-    const height = number('Height', 250);
-    const width = number('Width', 550);
-    const lineStroke = number('Stroke Width', 4);
-    const color = select('Color Scheme', schemes, 'cybertron');
-    const data = object('Data', multiDateData);
-
-    return (
-      <StackedAreaChart
-        width={width}
-        height={height}
-        series={
-          <StackedAreaSeries
-            colorScheme={color}
-            area={null}
-            line={<Line strokeWidth={lineStroke} />}
-          />
-        }
-        data={data}
-      />
-    );
-  })
-  .add('Stacked Normalized', () => {
-    const height = number('Height', 250);
-    const width = number('Width', 550);
-    const lineStroke = number('Stroke Width', 4);
-    const color = select('Color Scheme', schemes, 'cybertron');
-    const data = object('Data', multiDateData);
-
-    return (
-      <StackedNormalizedAreaChart
-        width={width}
-        height={height}
-        data={data}
-        series={
-          <StackedNormalizedAreaSeries
-            colorScheme={color}
-            area={null}
-            line={<Line strokeWidth={lineStroke} />}
-          />
-        }
-      />
-    );
-  });
-
-storiesOf('Charts/Line Chart/Gridlines', module)
-  .add('All Axes', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      gridlines={<GridlineSeries line={<Gridline direction="all" />} />}
-    />
-  ))
-  .add('X-Axis', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      gridlines={<GridlineSeries line={<Gridline direction="x" />} />}
-    />
-  ))
-  .add('Y-Axis', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      gridlines={<GridlineSeries line={<Gridline direction="y" />} />}
-    />
-  ));
-
-storiesOf('Charts/Line Chart/Circle Series', module)
-  .add('On', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      series={<LineSeries symbols={<PointSeries show={true} />} />}
-    />
-  ))
-  .add('Off', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      series={<LineSeries symbols={null} />}
-    />
-  ))
-  .add('On Hover', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      series={<LineSeries symbols={<PointSeries show="hover" />} />}
-    />
-  ))
-  .add('Only First', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      series={<LineSeries symbols={<PointSeries show="first" />} />}
-    />
-  ))
-  .add('Only Last', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      series={<LineSeries symbols={<PointSeries show="last" />} />}
-    />
-  ))
-  .add('Shapes', () => (
-    <LineChart
-      width={350}
-      height={250}
-      data={singleDateData}
-      series={
-        <LineSeries
-          symbols={
-            <PointSeries
-              show={true}
-              point={
-                <ScatterPoint
-                  symbol={() => {
-                    const d = symbol().type(symbolStar).size(175)();
-
-                    return (
-                      <path
-                        d={d!}
-                        style={{
-                          fill: 'lime',
-                          stroke: 'purple',
-                          strokeWidth: 1.5
-                        }}
-                      />
-                    );
-                  }}
-                />
-              }
-            />
-          }
-        />
-      }
-    />
-  ));
+export const LiveUpdating = () => <LiveUpdatingStory />;
 
 let interval;
 let offset = 0;
