+++ src/RadialAreaChart/RadialAreaChart.story.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
 import { RadialAreaChart } from './RadialAreaChart';
 import { medDateData, categoryData } from '../../demo';
 import { RadialAreaSeries, RadialArea } from './RadialAreaSeries';
@@ -15,77 +14,82 @@ import {
 import { schemes } from '../common/color';
 import { RadialGradient } from '../common/Gradient';
 
-storiesOf('Charts/Area Chart/Radial', module)
-  .add('Simple', () => {
-    const innerRadius = number('Inner Radius', 0.1);
-    const animated = boolean('Animated', true);
-    const hasGradient = boolean('Gradient', true);
-    const autoRotate = boolean('Auto Rotate Labels', true);
-    const color = select('Color Scheme', schemes, 'cybertron');
-    const gradient = hasGradient ? <RadialGradient /> : null;
-    const tickCount = number('Tick Count', 5);
-    const arcCount = number('Arc Count', 10);
-    const tickPosition = select(
-      'Tick Position',
-      {
-        inside: 'inside',
-        outside: 'outside'
-      },
-      'inside'
-    );
-    const interpolation = select(
-      'Interpolation',
-      {
-        linear: 'linear',
-        smooth: 'smooth'
-      },
-      'smooth'
-    );
-    const data = object('Data', medDateData);
+export default {
+  title: 'Charts/Area Chart/Radial'
+};
 
-    return (
-      <RadialAreaChart
-        height={450}
-        width={450}
-        data={data}
-        innerRadius={innerRadius}
-        series={
-          <RadialAreaSeries
-            colorScheme={color}
-            animated={animated}
-            interpolation={interpolation}
-            area={<RadialArea gradient={gradient} />}
-          />
-        }
-        axis={
-          <RadialAxis
-            arcs={<RadialAxisArcSeries count={arcCount} />}
-            ticks={
-              <RadialAxisTickSeries
-                count={tickCount}
-                tick={
-                  <RadialAxisTick
-                    line={<RadialAxisTickLine position={tickPosition} />}
-                    label={<RadialAxisTickLabel autoRotate={autoRotate} />}
-                  />
-                }
-              />
-            }
-          />
-        }
-      />
-    );
-  })
-  .add('Categorical Data', () => (
+export const Simple = () => {
+  const innerRadius = number('Inner Radius', 0.1);
+  const animated = boolean('Animated', true);
+  const hasGradient = boolean('Gradient', true);
+  const autoRotate = boolean('Auto Rotate Labels', true);
+  const color = select('Color Scheme', schemes, 'cybertron');
+  const gradient = hasGradient ? <RadialGradient /> : null;
+  const tickCount = number('Tick Count', 5);
+  const arcCount = number('Arc Count', 10);
+  const tickPosition = select(
+    'Tick Position',
+    {
+      inside: 'inside',
+      outside: 'outside'
+    },
+    'inside'
+  );
+  const interpolation = select(
+    'Interpolation',
+    {
+      linear: 'linear',
+      smooth: 'smooth'
+    },
+    'smooth'
+  );
+  const data = object('Data', medDateData);
+
+  return (
     <RadialAreaChart
-      data={categoryData}
-      height={300}
-      width={300}
-      axis={<RadialAxis type="category" />}
+      height={450}
+      width={450}
+      data={data}
+      innerRadius={innerRadius}
+      series={
+        <RadialAreaSeries
+          colorScheme={color}
+          animated={animated}
+          interpolation={interpolation}
+          area={<RadialArea gradient={gradient} />}
+        />
+      }
+      axis={
+        <RadialAxis
+          arcs={<RadialAxisArcSeries count={arcCount} />}
+          ticks={
+            <RadialAxisTickSeries
+              count={tickCount}
+              tick={
+                <RadialAxisTick
+                  line={<RadialAxisTickLine position={tickPosition} />}
+                  label={<RadialAxisTickLabel autoRotate={autoRotate} />}
+                />
+              }
+            />
+          }
+        />
+      }
     />
-  ))
-  .add('Resizable', () => (
-    <div style={{ width: '50vw', height: '75vh', border: 'solid 1px red' }}>
-      <RadialAreaChart data={medDateData} />
-    </div>
-  ));
+  );
+};
+
+export const CategoricalData = () => (
+  <RadialAreaChart
+    data={categoryData}
+    height={300}
+    width={300}
+    axis={<RadialAxis type="category" />}
+  />
+);
+
+export const Resizable = () => (
+  <div style={{ width: '50vw', height: '75vh', border: 'solid 1px red' }}>
+    <RadialAreaChart data={medDateData} />
+  </div>
+);
