+++ src/RadialAreaChart/RadialLineChart.story.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
 import { RadialAreaChart } from './RadialAreaChart';
 import { medDateData, categoryData } from '../../demo';
 import { RadialAreaSeries } from './RadialAreaSeries';
@@ -14,71 +13,75 @@ import {
 } from '../common/Axis';
 import { schemes } from '../common/color';
 
-storiesOf('Charts/Line Chart/Radial', module)
-  .add('Simple', () => {
-    const innerRadius = number('Inner Radius', 0.1);
-    const animated = boolean('Animated', true);
-    const color = select('Color Scheme', schemes, 'cybertron');
-    const autoRotate = boolean('Auto Rotate Labels', true);
-    const tickCount = number('Tick Count', 5);
-    const tickPosition = select(
-      'Tick Position',
-      {
-        inside: 'inside',
-        outside: 'outside'
-      },
-      'outside'
-    );
-    const arcCount = number('Arc Count', 10);
-    const interpolation = select(
-      'Interpolation',
-      {
-        linear: 'linear',
-        smooth: 'smooth'
-      },
-      'smooth'
-    );
-    const data = object('Data', medDateData);
+export default {
+  title: 'Charts/Line Chart/Radial'
+};
 
-    return (
-      <RadialAreaChart
-        height={450}
-        width={450}
-        data={data}
-        innerRadius={innerRadius}
-        series={
-          <RadialAreaSeries
-            area={null}
-            colorScheme={color}
-            animated={animated}
-            interpolation={interpolation}
-          />
-        }
-        axis={
-          <RadialAxis
-            arcs={<RadialAxisArcSeries count={arcCount} />}
-            ticks={
-              <RadialAxisTickSeries
-                count={tickCount}
-                tick={
-                  <RadialAxisTick
-                    line={<RadialAxisTickLine position={tickPosition} />}
-                    label={<RadialAxisTickLabel autoRotate={autoRotate} />}
-                  />
-                }
-              />
-            }
-          />
-        }
-      />
-    );
-  })
-  .add('Categorical Data', () => (
+export const Simple = () => {
+  const innerRadius = number('Inner Radius', 0.1);
+  const animated = boolean('Animated', true);
+  const color = select('Color Scheme', schemes, 'cybertron');
+  const autoRotate = boolean('Auto Rotate Labels', true);
+  const tickCount = number('Tick Count', 5);
+  const tickPosition = select(
+    'Tick Position',
+    {
+      inside: 'inside',
+      outside: 'outside'
+    },
+    'outside'
+  );
+  const arcCount = number('Arc Count', 10);
+  const interpolation = select(
+    'Interpolation',
+    {
+      linear: 'linear',
+      smooth: 'smooth'
+    },
+    'smooth'
+  );
+  const data = object('Data', medDateData);
+
+  return (
     <RadialAreaChart
-      data={categoryData}
-      height={300}
-      width={300}
-      series={<RadialAreaSeries area={null} />}
-      axis={<RadialAxis type="category" />}
+      height={450}
+      width={450}
+      data={data}
+      innerRadius={innerRadius}
+      series={
+        <RadialAreaSeries
+          area={null}
+          colorScheme={color}
+          animated={animated}
+          interpolation={interpolation}
+        />
+      }
+      axis={
+        <RadialAxis
+          arcs={<RadialAxisArcSeries count={arcCount} />}
+          ticks={
+            <RadialAxisTickSeries
+              count={tickCount}
+              tick={
+                <RadialAxisTick
+                  line={<RadialAxisTickLine position={tickPosition} />}
+                  label={<RadialAxisTickLabel autoRotate={autoRotate} />}
+                />
+              }
+            />
+          }
+        />
+      }
     />
-  ));
+  );
+};
+
+export const CategoricalData = () => (
+  <RadialAreaChart
+    data={categoryData}
+    height={300}
+    width={300}
+    series={<RadialAreaSeries area={null} />}
+    axis={<RadialAxis type="category" />}
+  />
+);
