+++ src/Sparkline/SparklineChart.story.tsx
@@ -1,4 +1,3 @@
-import { storiesOf } from '@storybook/react';
 import React from 'react';
 
 import {
@@ -13,17 +12,26 @@ import { AreaSparklineChart } from './AreaSparklineChart';
 import { BarSparklineChart } from './BarSparklineChart';
 import { SonarChart } from './SonarChart';
 
-storiesOf('Charts/Sparkline', module)
-  .add('Line', () => (
-    <SparklineChart width={200} height={55} data={medDateData} />
-  ))
-  .add('Area', () => (
-    <AreaSparklineChart width={200} height={85} data={singleDateData} />
-  ))
-  .add('Bar', () => (
-    <BarSparklineChart width={200} height={35} data={smallDateData} />
-  ))
-  .add('Sonar', () => <SonarChart width={300} height={50} data={sonarData} />)
-  .add('Sonar Empties', () => (
-    <SonarChart width={300} height={50} data={sonarDataEmpties} />
-  ));
+export default {
+  title: 'Charts/Sparkline'
+};
+
+export const Line = () => (
+  <SparklineChart width={200} height={55} data={medDateData} />
+);
+
+export const Area = () => (
+  <AreaSparklineChart width={200} height={85} data={singleDateData} />
+);
+
+export const Bar = () => (
+  <BarSparklineChart width={200} height={35} data={smallDateData} />
+);
+
+export const Sonar = () => (
+  <SonarChart width={300} height={50} data={sonarData} />
+);
+
+export const SonarEmpties = () => (
+  <SonarChart width={300} height={50} data={sonarDataEmpties} />
+);
